backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openpyxl
